move test case to test_tuner_utils.py
